refactor(service): #58 Tier 1.1 — split service.py into service/ mixin package
The 5,566-line LocalMacroDataService class is now seven domain-aligned mixin
files composed in service/__init__.py. Each method body is preserved
byte-for-byte; behavior is unchanged. The public import surface
(`from macro_data.service import LocalMacroDataService`) is identical, so
all downstream callers (scripts, tests, CLI, server) keep working without
edits.

Layout:
  service/
    __init__.py     LocalMacroDataService composes the six op mixins
    base.py         module helpers + LocalMacroDataServiceBase (init/invoke/_ensure_*)
    _calendar.py    68 _op_calendar_* + event_to_dict + parity helper      (4,207 LOC)
    _timeseries.py  23 indicator/catalog/concept/release/surprise/fed-comm   (626 LOC)
    _documents.py   list_items + get_document + subjects + backfill          (197 LOC)
    _news.py        recent/search/trends + live/article fetchers             (300 LOC)
    _market.py      snapshot + live-markets fetcher                           (53 LOC)
    _ops_health.py  refresh-all/run-schedule/source listing/health/alerts     (76 LOC)

Each mixin imports only the names its methods reference — no spurious
unused imports.

Verification: 1,673 unit tests pass on the branch (16 SDMX integration
failures are pre-existing flake on master too — network-bound tests).
Codex review confirmed the mixin split preserves method bodies and the
public import path.
